+++ Structurizr.Core/Model/ContainerInstance.cs
@@ -12,6 +12,9 @@ namespace Structurizr
     public sealed class ContainerInstance : Element
     {
 
+        private const int DefaultHealthCheckIntervalInSeconds = 60;
+        private const long DefaultHealthCheckTimeoutInMilliseconds = 0;
+
         public Container Container { get; internal set; }
 
         private string _containerId;
@@ -43,6 +46,22 @@ namespace Structurizr
             {
                 // no-op
             }
+        }
+
+        private HashSet<HttpHealthCheck> _healthChecks = new HashSet<HttpHealthCheck>();
+
+        /// <summary>
+        /// The set of health checks associated with this container instance.
+        /// </summary>
+        [DataMember(Name = "healthChecks", EmitDefaultValue = false)]
+        public HashSet<HttpHealthCheck> HealthChecks
+        {
+            get
+            {
+                return new HashSet<HttpHealthCheck>(_healthChecks);
+            }
+
+            internal set { _healthChecks = value; }
         }
 
         internal ContainerInstance() {
@@ -87,14 +106,70 @@ namespace Structurizr
         /// <param name="technology">the technology of the relationship</param>
         /// <returns>a Relationship object</returns>
         /// <exception cref="ArgumentException"></exception>
-        public Relationship Uses(ContainerInstance destination, string description, string technology) {
-            if (destination != null) {
+        public Relationship Uses(ContainerInstance destination, string description, string technology)
+        {
+            if (destination != null)
+            {
                 return Model.AddRelationship(this, destination, description, technology);
-            } else {
+            }
+            else
+            {
                 throw new ArgumentException("The destination of a relationship must be specified.");
             }
         }
 
+        /// <summary>
+        /// Adds a new health check, with the default interval (60 seconds) and timeout (0 milliseconds).
+        /// </summary>
+        /// <param name="name">The name of the health check.</param>
+        /// <param name="url">The URL of the health check.</param>
+        /// <returns>A HttpHealthCheck instance.</returns>
+        public HttpHealthCheck AddHealthCheck(string name, string url)
+        {
+            return AddHealthCheck(name, url, DefaultHealthCheckIntervalInSeconds, DefaultHealthCheckTimeoutInMilliseconds);
+        }
+
+        /// <summary>
+        /// Adds a new health check.
+        /// </summary>
+        /// <param name="name">The name of the health check.</param>
+        /// <param name="url">The URL of the health check.</param>
+        /// <param name="interval">The polling interval, in seconds.</param>
+        /// <param name="timeout">The timeout, in milliseconds.</param>
+        /// <returns>A HttpHealthCheck instance.</returns>
+        public HttpHealthCheck AddHealthCheck(string name, string url, int interval, long timeout)
+        {
+            if (name == null || name.Trim().Length == 0)
+            {
+                throw new ArgumentException("The name must not be null or empty.");
+            }
+
+            if (url == null || url.Trim().Length == 0)
+            {
+                throw new ArgumentException("The URL must not be null or empty.");
+            }
+
+            if (!Structurizr.Util.Url.IsUrl(url))
+            {
+                throw new ArgumentException(url + " is not a valid URL.");
+            }
+
+            if (interval < 0)
+            {
+                throw new ArgumentException("The polling interval must be zero or a positive integer.");
+            }
+
+            if (timeout < 0)
+            {
+                throw new ArgumentException("The timeout must be zero or a positive integer.");
+            }
+
+            HttpHealthCheck healthCheck = new HttpHealthCheck(name, url, interval, timeout);
+            _healthChecks.Add(healthCheck);
+
+            return healthCheck;
+        }
+
     }
-    
+
 }